Revert "Изменение порядка сортировки фильмов по лайкам (теперь по убыванию)"
This reverts commit 0b8a646091e97822fe226985c2ae29dc8520884f.

+++ src/main/java/ru/yandex/practicum/filmorate/controller/DirectorController.java
@@ -21,7 +21,7 @@ public class DirectorController {
     }
 
     @GetMapping("/{id}")
-    public Director findById(@PathVariable Integer id) {
+    public Director findById(@PathVariable Long id) {
         return directorService.findById(id);
     }
 
@@ -36,7 +36,7 @@ public class DirectorController {
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Integer id) {
+    public void delete(@PathVariable Long id) {
         directorService.delete(id);
     }
 }
